[MISC] Separate debug stream overload from aligned sequence concept.
This separation helps to avoid needless includes and dependencies for the aligned sequence concept.
